+++ tests/test-utils.php
@@ -14,8 +14,22 @@ class TrustedLoginUtilsTest extends WP_UnitTestCase {
 	 */
 	protected $wpdb;
 
+	/**
+	 * @var array
+	 */
+	protected $site_ids;
+
 	public function setUp(): void {
 		parent::setUp();
+
+		$this->create_sites();
+	}
+
+	public function create_sites() {
+		$this->site_ids = array();
+		$this->site_ids[] = get_current_blog_id();
+		$this->site_ids[] = $this->factory->blog->create( array( 'domain' => 'example.com', 'path' => '/' ) );
+		$this->site_ids[] = $this->factory->blog->create( array( 'domain' => 'example.com', 'path' => '/example/' ) );
 	}
 
 	public function testSetTransientWithoutExpiration() {
@@ -24,7 +38,7 @@ class TrustedLoginUtilsTest extends WP_UnitTestCase {
 
 		$result = Utils::set_transient( $transient, $value );
 
-		$this->assertTrue( $result );
+		$this->assertEquals( 1, $result );
 
 		$row = get_option( $transient );
 
@@ -42,7 +56,7 @@ class TrustedLoginUtilsTest extends WP_UnitTestCase {
 
 		$result = Utils::set_transient( $transient, $value, $expiration );
 
-		$this->assertTrue( $result );
+		$this->assertEquals( 1, $result );
 
 		$row = get_option( $transient );
 
@@ -58,7 +72,7 @@ class TrustedLoginUtilsTest extends WP_UnitTestCase {
 		$expiration_time = time() + $expiration;
 
 		$result = Utils::set_transient( $transient, $value, $expiration );
-		$this->assertTrue( $result );
+		$this->assertEquals( 1, $result );
 
 		$row = get_option( $transient );
 
@@ -75,6 +89,59 @@ class TrustedLoginUtilsTest extends WP_UnitTestCase {
 		$this->assertFalse( $result );
 	}
 
+	public function testSetTransientsWhenSwitchingSites() {
+		$transient = 'transient';
+		$value     = 'value';
+
+		// Initial site.
+		switch_to_blog( $this->site_ids[0] );
+
+		$result = Utils::set_transient( $transient, $value );
+		$this->assertEquals( 1, $result );
+
+		switch_to_blog( $this->site_ids[1] );
+
+		// Other sites should not have access to the transient.
+		$result = Utils::get_transient( $transient );
+		$this->assertFalse( $result );
+
+		$transient_site_2 = 'transient_site_2';
+		$value_site_2     = 'value_site_2';
+
+		$result = Utils::set_transient( $transient_site_2, $value_site_2 );
+		$this->assertEquals( 1, $result );
+
+		switch_to_blog( $this->site_ids[2] );
+
+		// Other sites should not have access to the transient.
+		$result = Utils::get_transient( $transient );
+		$this->assertFalse( $result );
+
+		$result = Utils::get_transient( $transient_site_2 );
+		$this->assertFalse( $result );
+
+		$transient_site_3 = 'transient_site_3';
+		$value_site_3     = 'value_site_3';
+
+		$result = Utils::set_transient( $transient_site_3, $value_site_3 );
+		$this->assertEquals( 1, $result );
+
+		switch_to_blog( $this->site_ids[0] );
+
+		// Sanity check that it's still there.
+		$result = Utils::get_transient( $transient );
+		$this->assertSame( $result, $value );
+
+		// And that the other site's transients are not.
+		$result = Utils::get_transient( $transient_site_2 );
+		$this->assertFalse( $result );
+
+		$result = Utils::get_transient( $transient_site_3 );
+		$this->assertFalse( $result );
+
+		restore_current_blog();
+	}
+
 	public function testSetTransientObject() {
 		$transient  = 'transient';
 		$value      = new stdClass();
@@ -82,7 +149,7 @@ class TrustedLoginUtilsTest extends WP_UnitTestCase {
 
 		$result = Utils::set_transient( $transient, $value );
 
-		$this->assertTrue( $result );
+		$this->assertEquals( 1, $result );
 
 		$row = get_option( $transient );
 
@@ -108,7 +175,7 @@ class TrustedLoginUtilsTest extends WP_UnitTestCase {
 
 		$result = Utils::set_transient( $transient, $value, $expiration );
 
-		$this->assertTrue( $result );
+		$this->assertEquals( 1, $result );
 
 		$row = Utils::get_transient( $transient );