+++ CHANGELOG.md
@@ -3,7 +3,10 @@
 ## 1.7.0 (January 29, 2024)
 
 - Added Utils class to handle common utility functions
-- Converted usage of `get_site_transient()` and `set_site_transient()` to using `Utils::get_site_transient()` and `Utils::set_site_transient()`. This prevents potential issues with object caching plugins that don't support transients, while allowing for auto-expiring data to be stored in the database.
+- Converted usage of `get_site_transient()` and `set_site_transient()` to using `Utils::get_transient()` and `Utils::set_transient()`.
+  - Scopes the storage to each blog instead of per-network, preventing potential issues with multisite
+  - Fixes potential issues with object caching plugins that don't support transients, while allowing for auto-expiring data to be stored in the database
+  - Prevents data from being "cleaned up" by site optimization plugins that remove expired transients
 
 ## 1.6.2 (January 26, 2024)